First changes

+++ dsd.go
@@ -10,12 +10,12 @@ var (
 
 // Decoder transforms encoded bits to float32 samples.
 type Decoder interface {
-	Decode(bits []byte) ([]float32, error)
+	Decode(bits []byte) ([]uint16, error)
 }
 
 // Encoder transforms float32 samples to encoded bits.
 type Encoder interface {
-	Encode([]float32) ([]byte, error)
+	Encode([]uint16) ([]byte, error)
 }
 
 // VoiceStream is a complete codec with a decode and an encoder.
